requirements.txt
algoliasearch==2.6.2
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.11
